requirements.txt
aiohttp>=3.9
pyvirtualcam>=0.4
numpy>=1.24
Pillow>=10.0
pystray>=0.19
qrcode>=7.4
cryptography>=41.0
